Contains the files for the 3D simulation of PhC defect waveguides. Please see documentation.pdf for more detail.
Loss engineering expansion for MPB

This is a free code for the computation of the ensemble averaged
propagation loss in photonic crystal waveguides.
It consists of several parts. First a control file for the free MIT
Photonics-Bands (MBP) package, written in scheme, to calculate the
electromagnetic modes and dispersion curve of the photonic crystal as
well as calculating the out-of-plane and Backscattering coefficients as
outlined in reference [1]. The additional files are for data analysis
and consits of linux shell and Matlab scripts.

All scripts can be downloaded under the free GPL license, no
installation of the extension is required, however an installed version
of MPB is required.
When publishing results obtained with this software, please reference
[1].

Features

Calculation of propagation loss for differen line defect waveguides
Prebuild support for dispersion engineering, using a variety of designs
[2]
Simultaneous calculation of band, structure, field distributions, group
index curve and propagation loss possible.
Free software under the GNU General Public License.

The calculation is time efficient and a 3D computation of the
bandstructure (30 bands), the group velocity curve, the electric and
magnetic field distribution, the distribution of |E|2 as well as the
out-of plane- and back-scattering coefficients for a slow light
waveguide on a Quad core desktop (2.3GHz) over 21k-points  using
16x16x30 grid is finished in approximately 3.5 hours.

Contact info

For support using this script or suggestion for improving it please
contact Daryl Beggs (beggs@amolf.nl) or Sebastian Schulz
(sas35@st-andrews.ac.uk), alternatively Sebastian Schulz also follows
discussions on the mpb-discuss mailing list.
For general problems or questions regarding MPB, please consult the MPB
Documentation (http://ab-initio.mit.edu/wiki/index.php/MPB_manual), the
mpb-discuss archives or the mpb-discuss mailing list.

References
[1] L.O'Faolain et al. Optics Express 18, 27627-27638 (2010)
[2] S.A.Schulz et al. Journal of Optics 12, 104004 (2010)